# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Penguin Statistics v3 API, re-designed to aim for lightweight on wire.",
        "title": "Penguin Statistics API",
        "contact": {
            "name": "AlvISs_Reimu",
            "url": "https://github.com/AlvISsReimu",
            "email": "alvissreimu@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "3.0.0-alpha.1"
    },
    "host": "https://penguin-stats.io",
    "basePath": "/api",
    "paths": {
        "/PenguinStats/api/v2/_private/result/matrix/{server}/{source}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Private"
                ],
                "summary": "Get DropMatrix",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server; default to CN",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Global or Personal; default to global",
                        "name": "source",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shims.DropMatrixQueryResult"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/PenguinStats/api/v2/_private/result/pattern/{server}/{source}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Private"
                ],
                "summary": "Get PatternMatrix",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server; default to CN",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Global or Personal; default to global",
                        "name": "source",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shims.PatternMatrixQueryResult"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/PenguinStats/api/v2/_private/result/trend/{server}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Private"
                ],
                "summary": "Get Trends",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server; default to CN",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shims.TrendQueryResult"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
...
        "/v3/items": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Get All Items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/models.Item"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "existence": {
                                                "$ref": "#/definitions/models.Existence"
                                            },
                                            "keywords": {
                                                "$ref": "#/definitions/models.Keywords"
                                            },
                                            "name": {
                                                "$ref": "#/definitions/models.I18nString"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/v3/items/{itemId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Get an Item with ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Item"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "existence": {
                                            "$ref": "#/definitions/models.Existence"
                                        },
                                        "keywords": {
                                            "$ref": "#/definitions/models.Keywords"
                                        },
                                        "name": {
                                            "$ref": "#/definitions/models.I18nString"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing itemId. Notice that this shall be the **string ID** of the item, instead of the internally used numerical ID of the item.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.Extras": {
            "type": "object",
            "additionalProperties": true
        },
        "errors.PenguinError": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "extras": {
                    "$ref": "#/definitions/errors.Extras"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request: some or all request parameters are invalid"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "models.Existence": {
            "type": "object",
            "required": [
                "CN",
                "JP",
                "KR",
                "US"
            ],
            "properties": {
                "CN": {
                    "description": "CN: 国服 Mainland China Server (maintained by Hypergryph Network Technology Co., Ltd.)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "JP": {
                    "description": "JP: 日服 Japan Server (maintained by Yostar Inc,.)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "KR": {
                    "description": "KR: 韩服 Korea Server (maintained by Yostar Limited)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "US": {
                    "description": "US: 美服/国际服 Global Server (maintained by Yostar Limited)",
                    "$ref": "#/definitions/models.ServerExistence"
                }
            }
        },
        "models.I18nOptionalString": {
            "type": "object",
            "properties": {
                "en": {
                    "description": "EN: English (en)",
                    "type": "string"
                },
                "ja": {
                    "description": "JP: 日本語 (ja)",
                    "type": "string"
                },
                "ko": {
                    "description": "KR: 한국어 (ko)",
                    "type": "string"
                },
                "zh": {
                    "description": "ZH: 中文 (zh-CN)",
                    "type": "string"
                }
            }
        },
        "models.I18nString": {
            "type": "object",
            "required": [
                "en",
                "ja",
                "ko",
                "zh"
            ],
            "properties": {
                "en": {
                    "description": "EN: English (en)",
                    "type": "string"
                },
                "ja": {
                    "description": "JP: 日本語 (ja)",
                    "type": "string"
                },
                "ko": {
                    "description": "KR: 한국어 (ko)",
                    "type": "string"
                },
                "zh": {
                    "description": "ZH: 中文 (zh-CN)",
                    "type": "string"
                }
            }
        },
        "models.Item": {
            "type": "object",
            "properties": {
                "existence": {
                    "description": "Existence is a map with server code as key and the existence of the item in that server as value.",
                    "type": "object"
                },
                "group": {
                    "description": "Group is an identifier of what the item actually is. For example, both orirock and orirock cube would have the same group, `orirock`.",
                    "type": "string"
                },
                "itemId": {
                    "description": "ArkItemID (itemId) is the previously used, string form ID of the item; in JSON-representation `itemId` is used as key.",
                    "type": "string"
                },
                "keywords": {
                    "description": "Keywords is an arbitrary JSON object containing the keywords of the item, for optimizing the results of the frontend built-in search engine.",
                    "type": "object"
                },
                "name": {
                    "description": "Name is a map with language code as key and the name of the item in that language as value.",
                    "type": "object"
                },
                "penguinItemId": {
                    "description": "ItemID (penguinItemId) is the numerical ID of the item.",
                    "type": "integer"
                },
                "rarity": {
                    "type": "integer"
                },
                "sortId": {
                    "description": "SortID is the sort position of the item.",
                    "type": "integer"
                },
                "sprite": {
                    "description": "Sprite describes the location of the item's sprite on the sprite image, in a form of Y:X.",
                    "type": "string"
                }
            }
        },
        "models.Keywords": {
            "type": "object",
            "properties": {
                "alias": {
                    "description": "Alias of the item,",
                    "$ref": "#/definitions/models.I18nOptionalString"
                },
                "pron": {
                    "description": "Pronounciation hints of the item",
                    "$ref": "#/definitions/models.I18nOptionalString"
                }
            }
        },
...
        "models.ServerExistence": {
            "type": "object",
            "required": [
                "exist"
            ],
            "properties": {
                "endTime": {
                    "type": "integer"
                },
                "exist": {
                    "type": "boolean"
                },
                "startTime": {
                    "type": "integer"
                }
            }
        },
...
        "null.Int": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "shims.DropMatrixQueryResult": {
            "type": "object",
            "properties": {
                "matrix": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shims.OneDropMatrixElement"
                    }
                }
            }
        },
...
        "shims.OneDrop": {
            "type": "object",
            "properties": {
                "itemId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "shims.OneDropMatrixElement": {
            "type": "object",
            "properties": {
                "end": {
                    "$ref": "#/definitions/null.Int"
                },
                "itemId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "stageId": {
                    "type": "string"
                },
                "start": {
                    "type": "integer"
                },
                "times": {
                    "type": "integer"
                }
            }
        },
        "shims.OneItemTrend": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "times": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "shims.OnePatternMatrixElement": {
            "type": "object",
            "properties": {
                "end": {
                    "$ref": "#/definitions/null.Int"
                },
                "pattern": {
                    "$ref": "#/definitions/shims.Pattern"
                },
                "quantity": {
                    "type": "integer"
                },
                "stageId": {
                    "type": "string"
                },
                "start": {
                    "type": "integer"
                },
                "times": {
                    "type": "integer"
                }
            }
        },
        "shims.Pattern": {
            "type": "object",
            "properties": {
                "drops": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shims.OneDrop"
                    }
                }
            }
        },
        "shims.PatternMatrixQueryResult": {
            "type": "object",
            "properties": {
                "pattern_matrix": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shims.OnePatternMatrixElement"
                    }
                }
            }
        },
...
        "shims.StageTrend": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/shims.OneItemTrend"
                    }
                },
                "startTime": {
                    "type": "integer"
                }
            }
        },
...
        "shims.TrendQueryResult": {
            "type": "object",
            "properties": {
                "trend": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/shims.StageTrend"
                    }
                }
            }
        },
...
    }
}